♻️  core: Split Field into a separate mod
